requirements.txt
Django~=3.1
djangorestframework
dj-rest-auth
django-allauth
